# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MeisterProPR.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/clients/{clientId}/prompt-overrides": {
      "get": {
        "tags": [
          "PromptOverrides"
        ],
        "summary": "Lists all prompt overrides for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of prompt overrides.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptOverrideDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Client not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PromptOverrides"
        ],
        "summary": "Creates a new prompt override for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Prompt override details.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prompt override created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptOverrideDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Client not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Override already exists for this scope/key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}/prompt-overrides/{id}": {
      "put": {
        "tags": [
          "PromptOverrides"
        ],
        "summary": "Replaces the override text of an existing prompt override.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Prompt override identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt override updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptOverrideDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Prompt override not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PromptOverrides"
        ],
        "summary": "Deletes a prompt override.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Prompt override identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Prompt override deleted."
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Prompt override not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PromptOverrideScopeDto": {
        "enum": [
          "clientScope",
          "crawlConfigScope"
        ],
        "type": "string",
        "description": "Scope of the prompt override. clientScope applies to all jobs for the client; crawlConfigScope applies only to jobs from the specified crawl configuration."
      },
      "PromptOverrideDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "crawlConfigId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scope": {
            "$ref": "#/components/schemas/PromptOverrideScopeDto"
          },
          "promptKey": {
            "type": "string",
            "description": "Named prompt segment: SystemPrompt, AgenticLoopGuidance, SynthesisSystemPrompt, QualityFilterSystemPrompt, PerFileContextPrompt"
          },
          "overrideText": {
            "type": "string",
            "description": "Full replacement text for the prompt segment."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "clientId",
          "scope",
          "promptKey",
          "overrideText",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false,
        "description": "A named prompt segment override for a client or crawl configuration."
      },
      "CreatePromptOverrideRequest": {
        "type": "object",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/PromptOverrideScopeDto"
          },
          "crawlConfigId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Required when scope is crawlConfigScope."
          },
          "promptKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "overrideText": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          }
        },
        "required": [
          "scope",
          "promptKey",
          "overrideText"
        ],
        "additionalProperties": false,
        "description": "Request body for creating a prompt override."
      },
      "UpdatePromptOverrideRequest": {
        "type": "object",
        "properties": {
          "overrideText": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          }
        },
        "required": [
          "overrideText"
        ],
        "additionalProperties": false,
        "description": "Request body for replacing the override text of an existing prompt override."
      }
    }
  },
...
}